chore: skip downloading / unzipping gameicons if file / folder already exists

+++ scripts/download.sh
@@ -1,5 +1,13 @@
 # download game icons (https://game-icons.net)
-wget -O ./iconpacks/game-icons-svg.zip https://game-icons.net/archives/svg/zip/000000/transparent/game-icons.net.svg.zip
-unzip ./iconpacks/game-icons-svg.zip
-mv ./icons/000000/transparent/1x1 ./iconpacks/gameicons
-rm -rf ./icons
+
+filepath="./iconpacks/game-icons-svg.zip"
+folderpath="./iconpacks/gameicons"
+
+if [ ! -d "$folderpath" ]; then
+    if [ ! -f "$filepath" ]; then
+        wget -O $filepath https://game-icons.net/archives/svg/zip/000000/transparent/game-icons.net.svg.zip
+    fi
+    unzip $filepath
+    mv ./icons/000000/transparent/1x1 $folderpath
+    rm -rf ./icons
+fi
